Update examples layouts. Separate ngrok config to file
