+++ README.md
@@ -61,17 +61,29 @@ ngrok.listen(server).then(() => {
 
 # Examples
 
+Degit can be used for cloning and running an example directory like this:
+```bash
+npx degit github:ngrok/ngrok-js/examples/<example> <folder-name>
+cd <folder-name>
+npm i
+```
+For example:
+```bash
+npx degit github:ngrok/ngrok-js/examples/express express && cd express && npm i
+```
+
+
 ## Frameworks
-* [Express](https://expressjs.com/) - [Quickstart Example](https://github.com/ngrok/ngrok-js/blob/main/examples/ngrok-express-quickstart.js), [Configuration Example](https://github.com/ngrok/ngrok-js/blob/main/examples/ngrok-express.js)
-* [Fastify](https://www.fastify.io/) - [Example](https://github.com/ngrok/ngrok-js/blob/main/examples/ngrok-fastify.js)
-* [Hapi](https://hapi.dev/) - [Example](https://github.com/ngrok/ngrok-js/blob/main/examples/ngrok-hapi.js)
-* [Koa](https://koajs.com/) - [Example](https://github.com/ngrok/ngrok-js/blob/main/examples/ngrok-koa.js)
+* [Express](https://expressjs.com/) - [Quickstart Example](https://github.com/ngrok/ngrok-js/blob/main/examples/express/ngrok-express-quickstart.js), [Configuration Example](https://github.com/ngrok/ngrok-js/blob/main/examples/express/ngrok-express.js)
+* [Fastify](https://www.fastify.io/) - [Example](https://github.com/ngrok/ngrok-js/blob/main/examples/fastify/ngrok-fastify.js)
+* [Hapi](https://hapi.dev/) - [Example](https://github.com/ngrok/ngrok-js/blob/main/examples/hapi/ngrok-hapi.js)
+* [Koa](https://koajs.com/) - [Example](https://github.com/ngrok/ngrok-js/blob/main/examples/koa/ngrok-koa.js)
 * [Nest.js](https://nestjs.com/) - [Example main.ts](https://github.com/ngrok/ngrok-js/blob/main/examples/nestjs/src/main.ts)
-* [Next.js](https://nextjs.org/) - [Example next.config.js](https://github.com/ngrok/ngrok-js/blob/main/examples/next.config.js)
-* [Remix](https://remix.run/) - [Example remix.config.js](https://github.com/ngrok/ngrok-js/blob/main/examples/remix/remix.config.js)
-* [Svelte](https://svelte.dev/) - [Example svelte.config.js](https://github.com/ngrok/ngrok-js/blob/main/examples/svelte/svelte.config.js) (works in vite.config.js too)
-* [Typescript](https://www.typescriptlang.org/) - [Example](https://github.com/ngrok/ngrok-js/blob/main/examples/ngrok-typescript.ts)
-* [Vue](https://vuejs.org/) - [Example vite.config.ts](https://github.com/ngrok/ngrok-js/blob/main/examples/vue/vite.config.ts)
+* [Next.js](https://nextjs.org/) - [Example next.config.js](https://github.com/ngrok/ngrok-js/blob/main/examples/nextjs/next.config.js) loading [ngrok.config.js](https://github.com/ngrok/ngrok-js/blob/main/examples/nextjs/ngrok.config.js)
+* [Remix](https://remix.run/) - [Example remix.config.js](https://github.com/ngrok/ngrok-js/blob/main/examples/remix/remix.config.js) loading [ngrok.config.js](https://github.com/ngrok/ngrok-js/blob/main/examples/remix/ngrok.config.js)
+* [Svelte](https://svelte.dev/) - [Example svelte.config.js](https://github.com/ngrok/ngrok-js/blob/main/examples/svelte/svelte.config.js) (works in vite.config.js too) loading [ngrok.config.cjs](https://github.com/ngrok/ngrok-js/blob/main/examples/svelte/ngrok.config.cjs)
+* [Typescript](https://www.typescriptlang.org/) - [Example ts-node](https://github.com/ngrok/ngrok-js/blob/main/examples/ngrok-typescript.ts)
+* [Vue](https://vuejs.org/) - [Example vite.config.ts](https://github.com/ngrok/ngrok-js/blob/main/examples/vue/vite.config.ts) loading [ngrok.config.ts](https://github.com/ngrok/ngrok-js/blob/main/examples/vue/ngrok.config.ts)
 * [Winston](https://github.com/winstonjs/winston#readme) Logging - [Example](https://github.com/ngrok/ngrok-js/blob/main/examples/ngrok-winston.js)
 
 ## Tunnel Types
