+++ examples/express/package.json
@@ -1,0 +1,7 @@
+{
+  "dependencies": {
+    "@ngrok/ngrok": "0.1.1",
+    "express": "^4.18.2"
+  },
+  "packageManager": "yarn@3.3.1"
+}
