+++ examples/hapi/package.json
@@ -1,0 +1,7 @@
+{
+  "devDependencies": {
+    "@ngrok/ngrok": "0.1.1",
+    "@hapi/hapi": "21.3.1"
+  },
+  "packageManager": "yarn@3.3.1"
+}
