+++ examples/koa/package.json
@@ -1,0 +1,7 @@
+{
+  "devDependencies": {
+    "@ngrok/ngrok": "0.1.1",
+    "koa": "2.14.2"
+  },
+  "packageManager": "yarn@3.3.1"
+}
