+++ examples/nestjs/src/main.ts
@@ -1,16 +1,17 @@
 import { NestFactory } from '@nestjs/core';
+import { Logger } from '@nestjs/common';
 import { AppModule } from './app.module';
 import { NgrokSessionBuilder } from '@ngrok/ngrok';
 
 async function bootstrap() {
-  // Added by ngrok
+  const port = 3000;
+  const app = await NestFactory.create(AppModule);
+  await app.listen(port);
+
+  // Setup ngrok ingress
   const session = await new NgrokSessionBuilder().authtokenFromEnv().connect();
   const tunnel = await session.httpEndpoint().listen();
-  console.log(`Ingress established at ${tunnel.url()}`);
-  tunnel.forwardTcp("localhost:3000");
-  // End added by ngrok
-
-  const app = await NestFactory.create(AppModule);
-  await app.listen(3000);
+  new Logger('main').log(`Ingress established at ${tunnel.url()}`);
+  tunnel.forwardTcp(`localhost:${port}`);
 }
 bootstrap();
