+++ examples/nextjs/next.config.js
@@ -1,0 +1,12 @@
+require('./ngrok.config.js');
+
+module.exports = async (phase, { defaultConfig }) => {
+  /**
+   * @type {import('next').NextConfig}
+   */
+  const nextConfig = {
+    /* config options here */
+  }
+
+  return nextConfig
+}
