+++ examples/remix/remix.config.js
@@ -1,14 +1,4 @@
-// setup ngrok ingress
-const ngrok = require('@ngrok/ngrok');
-var port = '3000';
-if (process.env.PORT) port = process.env.PORT;
-process.argv.forEach((item, index) => { if (item == '--port') port = process.argv[index+1]; });
-new ngrok.NgrokSessionBuilder().authtokenFromEnv().connect().then((session) => {
-  session.httpEndpoint().listen().then((tunnel) => {
-    console.log(`Forwarding to: localhost:${port} from ingress at: ${tunnel.url()}`);
-    tunnel.forwardTcp(`localhost:${port}`);
-  });
-});
+require('./ngrok.config.js');
 
 /** @type {import('@remix-run/dev').AppConfig} */
 module.exports = {
