+++ examples/svelte/svelte.config.js
@@ -1,19 +1,5 @@
 import adapter from '@sveltejs/adapter-auto';
-
-// setup ngrok ingress
-import ngrok from '@ngrok/ngrok';
-var host = 'localhost';
-var port = '5173';
-process.argv.forEach((item, index) => {
-  if (item == '--host') host = process.argv[index+1];
-  if (item == '--port') port = process.argv[index+1];
-});
-new ngrok.NgrokSessionBuilder().authtokenFromEnv().connect().then((session) => {
-  session.httpEndpoint().listen().then((tunnel) => {
-    console.log(`Forwarding to: ${host}:${port} from ingress at: ${tunnel.url()}`);
-    tunnel.forwardTcp(`${host}:${port}`);
-  });
-});
+import './ngrok.config.cjs'
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
